pyproject.toml
[project]
name = "elevenlabs-mcp"
version = "0.1.0"
description = "Eleven Labs MCP Server"
requires-python = ">=3.11,<3.13"
dependencies = [
    "mcp[cli]>=0.1.0",
    "fastapi==0.109.2",
    "uvicorn==0.27.1",
    "python-dotenv==1.0.1",
    "pydantic>=2.6.1",
    "httpx==0.28.1",
    "elevenlabs>=1.56.0",
    "python-magic>=0.4.27",
    "fuzzywuzzy==0.18.0",
    "python-Levenshtein>=0.25.0",
    "sounddevice==0.5.1",
    "soundfile==0.13.1"
]

[project.optional-dependencies]
dev = [
    "pre-commit==3.6.2",
    "ruff==0.3.0",
    "fastmcp==0.4.1",
    "pytest==8.0.0",
    "pytest-cov==4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["elevenlabs_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=elevenlabs_mcp --cov-report=term-missing"
